pyproject.toml
[project]
name = "sailplanedb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
]
